+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added command line argument for setting program number on Net Standard
  - Added features for getting the header from a core configuration
 
+### Changed
+ - Proxies are no longer instantiated on discovery and must be pre-configured
+
 ## [11.0.1] - 2018-11-20
 ### Changed
  - Fixed bug where USER migration would create redundant CommonConfig directory